refactor: auth 모듈 패키지 수정
